Version 6.9.11
 - onBackPress Когда удаляешь из ресВью выйти из режима, billList, search
